Move DefaultResponsiveSizes into a separate file
