+++ src/DefaultTheme.js
@@ -1,6 +1,6 @@
 /* eslint-disable key-spacing */
 
-export const DefaultTheme = {
+export default {
   // Colors
   background: '#F9F9F9',
   pageHeadingBackground: '#003B5C',
@@ -45,10 +45,3 @@ export const DefaultTheme = {
   // Modular scale ratio that is used to figure out all the different font sizes
   msRatio: 1.2
 };
-
-export const DefaultResponsiveSizes = [
-  {name: 'small', width: 360, height: 640},
-  {name: 'medium', width: 1024, height: 768},
-  {name: 'large', width: 1440, height: 900},
-  {name: 'xlarge', width: 1920, height: 1080}
-];
